subsys: Split monotonic cntrs in their own subsys
This splits the monotonic counter functionality
which previously existed in bl_storage in it's
own subsystem. Splitting this allows new users
to use the montotonic counters even if they
don't want to use the bl_storage subsys.

Ref: NCSDK-9045
