+++ subsys/CMakeLists.txt
@@ -12,6 +12,7 @@ add_subdirectory_ifdef(CONFIG_SECURE_BOOT bootloader/image)
 add_subdirectory_ifdef(CONFIG_SECURE_BOOT_CRYPTO bootloader/bl_crypto)
 add_subdirectory_ifdef(CONFIG_SECURE_BOOT_VALIDATION bootloader/bl_validation)
 add_subdirectory_ifdef(CONFIG_SECURE_BOOT_STORAGE bootloader/bl_storage)
+add_subdirectory_ifdef(CONFIG_SECURE_BOOT_STORAGE bootloader/bl_monotonic_counters)
 
 add_subdirectory(net)
 add_subdirectory_ifdef(CONFIG_ESB		esb)
