+++ subsys/bootloader/bl_monotonic_counters/CMakeLists.txt
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2020 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+zephyr_library()
+zephyr_library_sources(bl_monotonic_counters.c)
